requirements.txt
requests==2.31.0
python-dotenv==1.0.0
gspread==6.0.0
pandas==2.2.0
rich==13.7.0
